linting and watermark
